Resolve a number of egs_view compiler warnings
Also ran astyle

+++ HEN_HOUSE/egs++/egs_ausgab_object.cpp
@@ -77,6 +77,6 @@ int EGS_AusgabObject::nObjects() {
     return ausgab_object_creator.nObjects();
 }
 
-EGS_AusgabObject *EGS_AusgabObject::getObject(int j) {
+EGS_AusgabObject *EGS_AusgabObject::getObject(size_t j) {
     return (EGS_AusgabObject *) ausgab_object_creator.getObject(j);
 }
